requirements.txt
streamlit==1.37.0
pandas==2.2.2
chromadb==0.5.5
PyPDF2==3.0.1
PyYAML==6.0.2
sentence-transformers==3.0.1
numpy==1.26.4
openai>=1.40.0,<2
